% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mlapp.R
\name{withConsoleRedirect}
\alias{withConsoleRedirect}
\title{withConsoleRedirect}
\usage{
withConsoleRedirect(containerId, expr)
}
\arguments{
\item{containerId}{the ui container id of the console to print the text}

\item{expr}{the actual text}
}
\value{

}
\description{
withConsoleRedirect
}
